Move color fixtures to top-level conftest
